Extract html report to own module
