+++ detekt-cli/build.gradle.kts
@@ -5,12 +5,10 @@ application {
 dependencies {
     implementation(kotlin("stdlib-jdk8"))
     implementation(project(":detekt-core"))
+    implementation(project(":detekt-report-html"))
     implementation(project(":detekt-report-xml"))
     runtimeOnly(project(":detekt-rules"))
     implementation("com.beust:jcommander:${Versions.JCOMMANDER}")
-    implementation("org.jetbrains.kotlinx:kotlinx-html-jvm:${Versions.KOTLINX_HTML}") {
-        exclude(group = "org.jetbrains.kotlin")
-    }
 
     testImplementation(project(":detekt-test"))
     testImplementation(project(":detekt-rules"))
