+++ detekt-cli/src/main/resources/META-INF/services/io.gitlab.arturbosch.detekt.api.OutputReport
@@ -1,2 +1,1 @@
 io.gitlab.arturbosch.detekt.cli.out.TxtOutputReport
-io.gitlab.arturbosch.detekt.cli.out.HtmlOutputReport
