+++ detekt-report-html/src/main/resources/default-html-report-template.html
