+++ detekt-report-html/src/test/resources/HtmlOutputFormatTest.html
