+++ detekt-report-html/src/test/resources/META-INF/MANIFEST.MF
@@ -1,0 +1,1 @@
+DetektVersion: 1.9.1
