+++ docs/pages/kdoc/detekt-api/io.gitlab.arturbosch.detekt.api.internal/which-detekt.md
@@ -1,0 +1,12 @@
+---
+title: whichDetekt - detekt-api
+---
+
+[detekt-api](../index.html) / [io.gitlab.arturbosch.detekt.api.internal](index.html) / [whichDetekt](./which-detekt.html)
+
+# whichDetekt
+
+`fun whichDetekt(): `[`String`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-string/index.html)`?`
+
+Returns the bundled detekt version.
+
