+++ docs/pages/kdoc/detekt-api/io.gitlab.arturbosch.detekt.api.internal/which-java.md
@@ -1,0 +1,12 @@
+---
+title: whichJava - detekt-api
+---
+
+[detekt-api](../index.html) / [io.gitlab.arturbosch.detekt.api.internal](index.html) / [whichJava](./which-java.html)
+
+# whichJava
+
+`fun whichJava(): `[`String`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-string/index.html)
+
+Returns the version of the running JVM.
+
