+++ docs/pages/kdoc/detekt-api/io.gitlab.arturbosch.detekt.api.internal/which-o-s.md
@@ -1,0 +1,12 @@
+---
+title: whichOS - detekt-api
+---
+
+[detekt-api](../index.html) / [io.gitlab.arturbosch.detekt.api.internal](index.html) / [whichOS](./which-o-s.html)
+
+# whichOS
+
+`fun whichOS(): `[`String`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-string/index.html)
+
+Returns the name of the running OS.
+
